refactor: remove telemetry service index exports
