+++ packages/salesforcedx-utils-vscode/src/telemetry/index.ts
@@ -5,4 +5,13 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-export { TelemetryService, TelemetryBuilder, TelemetryData } from './telemetry';
+/**
+ * This file is intentionally left blank. Telemetry Service is exported
+ * in the root of salesforcedx-utils-vscode.
+ *
+ * If your service or util is dependent on telemetry service, please leave the index
+ * of your service folder empty as well, and export your service in src/index.ts.
+ * Otherwise, you will have a duplicate copy of telemetry service in your folder,
+ * for webpack treats the index of each folder as an entry point
+ * and bundles all of its dependencies.
+ */
